refactor: small ref
