+++ src/components/Layout/Header/index.tsx
@@ -3,7 +3,7 @@ import generalClasses from '@src/styles/general.module.scss';
 import { useAuthState } from 'react-firebase-hooks/auth';
 import { useTranslation } from 'react-i18next';
 import { NavLink } from 'react-router-dom';
-import { auth, logout } from '@src/firebase';
+import { auth, logout } from '@src/shared/api/firebase';
 import LanguageSelector from '@src/components/LanguageSelector';
 
 const Header = () => {
@@ -22,34 +22,34 @@ const Header = () => {
                   {t('mainPage')}
                 </NavLink>
               </li>
-              <li>
-                {user && (
+              {user && (
+                <li>
                   <NavLink to="/graphiql" className={classes.menuLink}>
                     {t('graphiql')}
                   </NavLink>
-                )}
-              </li>
-              <li>
-                {!user && (
+                </li>
+              )}
+              {!user && (
+                <li>
                   <NavLink to="/auth" className={classes.menuLink}>
                     {t('signIn')}
                   </NavLink>
-                )}
-              </li>
-              <li>
-                {!user && (
+                </li>
+              )}
+              {!user && (
+                <li>
                   <NavLink to="/register" className={classes.menuLink}>
                     {t('signUp')}
                   </NavLink>
-                )}
-              </li>
-              <li>
-                {user && (
+                </li>
+              )}
+              {user && (
+                <li>
                   <button onClick={logout} className={classes.menuLink}>
                     {t('logOut')}
                   </button>
-                )}
-              </li>
+                </li>
+              )}
             </ul>
           </nav>
           <div className={classes.toolbar}>
